BookTableBuddy/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
cssselect==1.3.0
cssutils==2.11.1
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.6.0
idna==3.10
lxml==5.4.0
MarkupSafe==3.0.2
more-itertools==10.7.0
multidict==6.4.3
pillow==11.2.1
premailer==3.10.0
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-http-client==3.3.7
requests==2.32.3
sendgrid==6.12.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
twilio==9.6.0
typing_extensions==4.13.1
urllib3==2.4.0
Werkzeug==2.3.8
yagmail==0.15.293
yarl==1.20.0
